+++ libs/langchain/langchain/output_parsers/__init__.py
@@ -34,7 +34,7 @@ from langchain.output_parsers.regex import RegexParser
 from langchain.output_parsers.regex_dict import RegexDictParser
 from langchain.output_parsers.retry import RetryOutputParser, RetryWithErrorOutputParser
 from langchain.output_parsers.structured import ResponseSchema, StructuredOutputParser
-from langchain.output_parsers.xml import XMLOutputParser
+from langchain.output_parsers.xml_parser import XMLOutputParser
 
 __all__ = [
     "BooleanOutputParser",
